// src/app/models/invoice.ts
export interface Invoice{
    id_invoice_cab?:    number;
    ord_comp_cab_id?:   number;
    num_orden_cab?:     string;
    num_consigned_cab?: string;
    fecha_inv_cab?:     string;
    nomb_contact_cab?:  string;
    telef_contact_cab?: string;
    country_cab?:       string;
    state_cab?:         string;
    terms_pay_cab?:     string;
}